--- Card.h
@@ -1,12 +1,0 @@
-#pragma once
-#include <iostream>
-class Card
-{
-private:
-	bool isWinner = false;
-public:
-	int cardIndex = 0;
-	void SetAsWinner(bool _value);
-	bool IsWinner();
-};
-
