+++ Debug/MonteCarloGame.Build.CppClean.log
@@ -1,0 +1,14 @@
+d:\misc projects\vs cpp\montecarlogame\debug\vc142.pdb
+d:\misc projects\vs cpp\montecarlogame\debug\vc142.idb
+d:\misc projects\vs cpp\montecarlogame\debug\montecarlogame.obj
+d:\misc projects\vs cpp\montecarlogame\debug\carddeck.obj
+d:\misc projects\vs cpp\montecarlogame\debug\card.obj
+d:\misc projects\vs cpp\montecarlogame\debug\montecarlogame.ilk
+d:\misc projects\vs cpp\montecarlogame\debug\montecarlogame.exe
+d:\misc projects\vs cpp\montecarlogame\debug\montecarlogame.pdb
+d:\misc projects\vs cpp\montecarlogame\debug\montecarlogame.tlog\cl.command.1.tlog
+d:\misc projects\vs cpp\montecarlogame\debug\montecarlogame.tlog\cl.read.1.tlog
+d:\misc projects\vs cpp\montecarlogame\debug\montecarlogame.tlog\cl.write.1.tlog
+d:\misc projects\vs cpp\montecarlogame\debug\montecarlogame.tlog\link.command.1.tlog
+d:\misc projects\vs cpp\montecarlogame\debug\montecarlogame.tlog\link.read.1.tlog
+d:\misc projects\vs cpp\montecarlogame\debug\montecarlogame.tlog\link.write.1.tlog
