+++ src/Card.h
@@ -1,0 +1,12 @@
+#pragma once
+#include <iostream>
+class Card
+{
+private:
+	bool isWinner = false;
+public:
+	int cardIndex = 0;
+	void SetAsWinner(bool _value);
+	bool IsWinner();
+};
+
